requirements.txt

streamlit
openai
groq
pymupdf
python-docx
python-dotenv

